Move TargetTechnology and TargetVulnerability to InputTechnology and InputVulnerability in parameters module
