+++ rekono/parameters/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from parameters.models import InputTechnology, InputVulnerability
+
+# Register your models here.
+
+admin.site.register(InputTechnology)
+admin.site.register(InputVulnerability)
