+++ rekono/parameters/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ParametersConfig(AppConfig):
+    '''Parameters Django application.'''
+
+    name = 'parameters'
