+++ rekono/parameters/serializers.py
@@ -1,0 +1,69 @@
+from typing import Any, Dict
+
+from django.forms import ValidationError
+from rest_framework import serializers
+
+from parameters.models import InputTechnology, InputVulnerability
+
+
+class InputTechnologySerializer(serializers.ModelSerializer):
+    '''Serializer to manage input technologies via API.'''
+
+    class Meta:
+        '''Serializer metadata.'''
+
+        model = InputTechnology
+        # Input technology fields exposed via API
+        fields = ('id', 'target_port', 'name', 'version')
+
+    def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
+        '''Validate the provided data before use it.
+
+        Args:
+            attrs (Dict[str, Any]): Provided data
+
+        Raises:
+            ValidationError: Raised if provided data is invalid
+
+        Returns:
+            Dict[str, Any]: Data after validation process
+        '''
+        attrs = super().validate(attrs)
+        if InputTechnology.objects.filter(
+            target_port=attrs['target_port'],
+            name=attrs['name'],
+            version=attrs['version']
+        ).exists():
+            raise ValidationError({
+                'name': 'This name already exists in this target port',
+                'version': 'This version already exists for this technology in this target port'
+            })
+        return attrs
+
+
+class InputVulnerabilitySerializer(serializers.ModelSerializer):
+    '''Serializer to manage input vulnerabilities via API.'''
+
+    class Meta:
+        '''Serializer metadata.'''
+
+        model = InputVulnerability
+        # Input vulnerabilities fields exposed via API
+        fields = ('id', 'target_port', 'cve')
+
+    def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
+        '''Validate the provided data before use it.
+
+        Args:
+            attrs (Dict[str, Any]): Provided data
+
+        Raises:
+            ValidationError: Raised if provided data is invalid
+
+        Returns:
+            Dict[str, Any]: Data after validation process
+        '''
+        attrs = super().validate(attrs)
+        if InputVulnerability.objects.filter(target_port=attrs['target_port'], cve=attrs['cve']).exists():
+            raise ValidationError({'cve': 'This CVE already exists in this target port'})
+        return attrs
