+++ rekono/parameters/urls.py
@@ -1,0 +1,9 @@
+from rest_framework.routers import SimpleRouter
+
+from parameters.views import InputTechnologyViewSet, InputVulnerabilityViewSet
+
+router = SimpleRouter()
+router.register('parameters/technologies', InputTechnologyViewSet)
+router.register('parameters/technologies', InputVulnerabilityViewSet)
+
+urlpatterns = router.urls
