+++ rekono/targets/admin.py
@@ -1,11 +1,8 @@
 from django.contrib import admin
 
-from targets.models import (Target, TargetPort, TargetTechnology,
-                            TargetVulnerability)
+from targets.models import Target, TargetPort
 
 # Register your models here.
 
 admin.site.register(Target)
 admin.site.register(TargetPort)
-admin.site.register(TargetTechnology)
-admin.site.register(TargetVulnerability)
