+++ rekono/targets/filters.py
@@ -1,8 +1,7 @@
 from django_filters import rest_framework
 from django_filters.rest_framework.filters import OrderingFilter
 
-from targets.models import (Target, TargetPort, TargetTechnology,
-                            TargetVulnerability)
+from targets.models import Target, TargetPort
 
 
 class TargetFilter(rest_framework.FilterSet):
@@ -42,51 +41,4 @@ class TargetPortFilter(rest_framework.FilterSet):
             'target__target': ['exact', 'icontains'],
             'target__type': ['exact'],
             'port': ['exact']
-        }
-
-
-class TargetTechnologyFilter(rest_framework.FilterSet):
-    '''FilterSet to filter and sort Target Technology entities.'''
-
-    o = OrderingFilter(fields=('target_port', 'name'))                          # Ordering fields
-
-    class Meta:
-        '''FilterSet metadata.'''
-
-        model = TargetTechnology
-        fields = {                                                              # Filter fields
-            'target_port': ['exact'],
-            'target_port__port': ['exact'],
-            'target_port__target': ['exact'],
-            'target_port__target__project': ['exact'],
-            'target_port__target__project__name': ['exact', 'icontains'],
-            'target_port__target__project__owner': ['exact'],
-            'target_port__target__project__owner__username': ['exact', 'icontains'],
-            'target_port__target__target': ['exact', 'icontains'],
-            'target_port__target__type': ['exact'],
-            'name': ['exact', 'icontains'],
-            'version': ['exact', 'icontains'],
-        }
-
-
-class TargetVulnerabilityFilter(rest_framework.FilterSet):
-    '''FilterSet to filter and sort Target Vulnerability entities.'''
-
-    o = OrderingFilter(fields=('target_port', 'cve'))                           # Ordering fields
-
-    class Meta:
-        '''FilterSet metadata.'''
-
-        model = TargetVulnerability
-        fields = {                                                              # Filter fields
-            'target_port': ['exact'],
-            'target_port__port': ['exact'],
-            'target_port__target': ['exact'],
-            'target_port__target__project': ['exact'],
-            'target_port__target__project__name': ['exact', 'icontains'],
-            'target_port__target__project__owner': ['exact'],
-            'target_port__target__project__owner__username': ['exact', 'icontains'],
-            'target_port__target__target': ['exact', 'icontains'],
-            'target_port__target__type': ['exact'],
-            'cve': ['exact']
         }
