+++ rekono/targets/serializers.py
@@ -3,47 +3,22 @@ from typing import Any, Dict
 from django.forms import ValidationError
 from rest_framework import serializers
 
-from targets.models import (Target, TargetPort, TargetTechnology,
-                            TargetVulnerability)
+from targets.models import Target, TargetPort
 from targets.utils import get_target_type
 
 
-class TargetPortSerializer(serializers.ModelSerializer):
-    '''Serializer to manage target ports via API.'''
+class SimplyTargetSerializer(serializers.ModelSerializer):
+    '''Simply serializer to include target main data in other serializers.'''
 
     class Meta:
         '''Serializer metadata.'''
 
-        model = TargetPort
-        fields = (                                                              # Target port fields exposed via API
-            'id', 'target', 'port', 'target_technologies', 'target_vulnerabilities', 'authentication'
-        )
-        # Read only fields
-        read_only_fields = ('target_technologies', 'target_vulnerabilities', 'authentication')
-
-    def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
-        '''Validate the provided data before use it.
-
-        Args:
-            attrs (Dict[str, Any]): Provided data
-
-        Raises:
-            ValidationError: Raised if provided data is invalid
-
-        Returns:
-            Dict[str, Any]: Data after validation process
-        '''
-        attrs = super().validate(attrs)
-        if TargetPort.objects.filter(target=attrs['target'], port=attrs['port']).exists():
-            raise ValidationError({'port': 'This port already exists in this target'})
-        return attrs
+        model = Target
+        fields = ('id', 'project', 'target', 'type')                            # Target fields exposed via API
 
 
 class TargetSerializer(serializers.ModelSerializer):
     '''Serializer to manage targets via API.'''
-
-    # Target ports details for read operations
-    target_ports = TargetPortSerializer(many=True, read_only=True)
 
     class Meta:
         '''Serializer metadata.'''
@@ -73,25 +48,18 @@ class TargetSerializer(serializers.ModelSerializer):
         return attrs
 
 
-class SimplyTargetSerializer(serializers.ModelSerializer):
-    '''Simply serializer to include target main data in other serializers.'''
+class TargetPortSerializer(serializers.ModelSerializer):
+    '''Serializer to manage target ports via API.'''
 
     class Meta:
         '''Serializer metadata.'''
 
-        model = Target
-        fields = ('id', 'project', 'target', 'type')                            # Target fields exposed via API
-
-
-class TargetVulnerabilitySerializer(serializers.ModelSerializer):
-    '''Serializer to manage target vulnerabilities via API.'''
-
-    class Meta:
-        '''Serializer metadata.'''
-
-        model = TargetVulnerability
-        # Target vulnerabilities fields exposed via API
-        fields = ('id', 'target_port', 'cve')
+        model = TargetPort
+        fields = (                                                              # Target port fields exposed via API
+            'id', 'target', 'port', 'input_technologies', 'inputt_vulnerabilities', 'authentication'
+        )
+        # Read only fields
+        read_only_fields = ('input_technologies', 'inputt_vulnerabilities', 'authentication')
 
     def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
         '''Validate the provided data before use it.
@@ -106,41 +74,6 @@ class TargetVulnerabilitySerializer(serializers.ModelSerializer):
             Dict[str, Any]: Data after validation process
         '''
         attrs = super().validate(attrs)
-        if TargetVulnerability.objects.filter(target_port=attrs['target_port'], cve=attrs['cve']).exists():
-            raise ValidationError({'cve': 'This CVE already exists in this target port'})
-        return attrs
-
-
-class TargetTechnologySerializer(serializers.ModelSerializer):
-    '''Serializer to manage target technologies via API.'''
-
-    class Meta:
-        '''Serializer metadata.'''
-
-        model = TargetTechnology
-        # Target technology fields exposed via API
-        fields = ('id', 'target_port', 'name', 'version')
-
-    def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
-        '''Validate the provided data before use it.
-
-        Args:
-            attrs (Dict[str, Any]): Provided data
-
-        Raises:
-            ValidationError: Raised if provided data is invalid
-
-        Returns:
-            Dict[str, Any]: Data after validation process
-        '''
-        attrs = super().validate(attrs)
-        if TargetTechnology.objects.filter(
-            target_port=attrs['target_port'],
-            name=attrs['name'],
-            version=attrs['version']
-        ).exists():
-            raise ValidationError({
-                'name': 'This name already exists in this target port',
-                'version': 'This version already exists for this technology in this target port'
-            })
+        if TargetPort.objects.filter(target=attrs['target'], port=attrs['port']).exists():
+            raise ValidationError({'port': 'This port already exists in this target'})
         return attrs
